+++ docs/changelog.md
@@ -19,7 +19,9 @@ MML changes:
     * Positive `px` values pan to the right. `px+64` is 100% to the right.
  * Out-of-range errors now show the value that was out-of-range
  * Added subroutine no-instrument note tracking.
-   The compiler now emits an error if a subroutine call would play an out-of-range note.
+   * The compiler now emits an error if a subroutine call would play an out-of-range note.
+   * The GUI's subroutine playback will now use the first MML instrument that can play all of the
+     subroutine's no-instrument notes.
 
 BRR changes:
  * Added gaussian overflow glitch detector (3 maximum-negative values in a row)